GUI/requirements.txt
pygame>=2.0.0
paho-mqtt>=1.6.0
